Create Staircase
Staircase detail
This is a staircase of size :
   #
  ##
 ###
####
Its base and height are both equal to n. It is drawn using # symbols and spaces. Any spaces do not precede the last line.
Write a program that prints a staircase of size.
Function Description
Complete the staircase function in the editor below.
staircase has the following parameter(s):
Print a staircase as described above.